@@ -1,6 +1,7 @@
 package com.example.tolearn;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.view.Window;
 import android.view.WindowManager;
